Moved windows specific Implementation and renamed mincore to kernal32
